Fix: circular reference error
